updated Jenkinsfile

+++ Jenkinsfile
@@ -9,7 +9,10 @@ pipeline{
             steps{
                 
                 //remove previously added repo
-                //sh 'rm -rf voting-app*'
+                if [ -d "/var/lib/jenkins/workspace/voting-app" ]
+                then
+                sh 'rm -rf /var/lib/jenkins/workspace/voting-app*'
+                fi
 
                 //repository checkout
                 sh 'git clone https://github.com/SyedYakhub/voting-app.git'
@@ -24,8 +27,11 @@ pipeline{
                 sh "docker login -u yakhub4881 -p ${DockerPasswd}"
 
                 //remove exisiting docker images
+                if [ -n "$(docker images --format '{{.Repository}}:{{.Tag}}' | grep 'yakhub4881/$JOB_NAME-vote')" ] || [ -n "$(docker images --format '{{.Repository}}:{{.Tag}}' | grep 'yakhub4881/$JOB_NAME-vote')" ]
+                then
                 sh "docker rmi -f \$(docker images --format '{{.Repository}}:{{.Tag}}' | grep '$JOB_NAME-vote')"
                 sh "docker rmi -f \$(docker images --format '{{.Repository}}:{{.Tag}}' | grep 'yakhub4881/$JOB_NAME-vote')"
+                fi
 
                 
                 //voter
